refactored load_areas to allow arbitrary zip/shape file from url
