+++ api/app/signals/apps/dataset/sources/__init__.py
@@ -1,9 +1,11 @@
 from signals.apps.dataset.sources.cbs import CBSBoundariesLoader
 from signals.apps.dataset.sources.gebieden import APIGebiedenLoader
+from signals.apps.dataset.sources.shape import ShapeBoundariesLoader
 from signals.apps.dataset.sources.sia_stadsdeel import SIAStadsdeelLoader
 
 __all__ = [
     'APIGebiedenLoader',
     'SIAStadsdeelLoader',
     'CBSBoundariesLoader',
+    'ShapeBoundariesLoader',
 ]
