+++ api/app/signals/apps/dataset/sources/gebieden.py
@@ -47,7 +47,8 @@ class APIGebiedenLoader(AreaLoader):
     }
     PROVIDES = ['stadsdeel', 'buurt', 'wijk']
 
-    def __init__(self, type_string, directory=None):
+    def __init__(self, **options):
+        type_string = options['type_string']
         assert type_string in self.CODE_FIELDS
 
         self.area_type, _ = AreaType.objects.get_or_create(
